fix: guard against partial mutation when a substitution's incoming player is unknown
Copilot review on #80: automaticSubstitution/manualSubstitution each make
two transition() calls with no rollback. If the incoming player's id had
no matching Player in the context, the second call could throw after the
first (benching the outgoing player) had already succeeded, leaving the
game with no active replacement.

Both now validate the incoming player exists before making either
transition() call, so the swap either fully succeeds or doesn't mutate
game state at all. Adds regression tests for both.

Also splits GameManagerTests.swift into one file per suite
(GameManagerTransitionTests/SubstitutionTests/AdHocAndDurationTests) to
stay under SwiftLint's file_length after the new tests pushed it over 400
lines.
